+++ app/src/main/java/com/example/twistedbets/models/match/Match.kt
@@ -2,9 +2,8 @@ package com.example.twistedbets.models.match
 
 import com.example.twistedbets.models.match.team.Team
 import com.google.gson.annotations.SerializedName
-import java.math.BigInteger
 
-class Match(    @SerializedName("gameId") var gameId: BigInteger,
+class Match(    @SerializedName("gameId") var gameId: Long,
                 @SerializedName("platformId") var platformId: String,
                 @SerializedName("gameDuration") var gameDuration: Int,
                 @SerializedName("gameMode") var gameMode: String,
