+++ app/src/main/java/com/example/twistedbets/models/match/MatchListItem.kt
@@ -1,17 +1,18 @@
 package com.example.twistedbets.models.match
 
 import com.google.gson.annotations.SerializedName
-import java.math.BigInteger
 
 data class MatchListItem(
 
     @SerializedName("platformId") var platformId: String,
-    @SerializedName("gameId") var gameId: BigInteger,
+    @SerializedName("gameId") var gameId: Long,
     @SerializedName("champion") var champion: Int,
     @SerializedName("queue") var queue: Int,
     @SerializedName("season") var season: Int,
-    @SerializedName("timestamp") var timestamp: BigInteger,
+    @SerializedName("timestamp") var timestamp: Long,
     @SerializedName("role") val role : String,
     @SerializedName("lane") val lane : String
+
+
 ) {
 }
